+++ tests/camel-kamelets-itest/src/test/resources/aws/s3/amazonS3Client.groovy
@@ -15,25 +15,31 @@
  * limitations under the License.
  */
 
-package aws.s3
 
+
+import org.apache.camel.CamelContext
 import software.amazon.awssdk.auth.credentials.AwsBasicCredentials
 import software.amazon.awssdk.auth.credentials.StaticCredentialsProvider
 import software.amazon.awssdk.regions.Region
 import software.amazon.awssdk.services.s3.S3Client
 
+if (context.getReferenceResolver().isResolvable(CamelContext.class)) {
+    println "Destroying former S3 client instance"
+    context.getReferenceResolver().resolve(CamelContext.class).getRegistry().unbind("amazonS3Client")
+}
+
 S3Client s3 = S3Client
         .builder()
-        .endpointOverride(URI.create("${CITRUS_TESTCONTAINERS_LOCALSTACK_SERVICE_URL}"))
+        .endpointOverride(URI.create('${CITRUS_TESTCONTAINERS_LOCALSTACK_SERVICE_URL}'))
         .credentialsProvider(StaticCredentialsProvider.create(
                 AwsBasicCredentials.create(
-                        "${CITRUS_TESTCONTAINERS_LOCALSTACK_ACCESS_KEY}",
-                        "${CITRUS_TESTCONTAINERS_LOCALSTACK_SECRET_KEY}")
+                        '${CITRUS_TESTCONTAINERS_LOCALSTACK_ACCESS_KEY}',
+                        '${CITRUS_TESTCONTAINERS_LOCALSTACK_SECRET_KEY}')
         ))
         .forcePathStyle(true)
-        .region(Region.of("${CITRUS_TESTCONTAINERS_LOCALSTACK_REGION}"))
+        .region(Region.of('${CITRUS_TESTCONTAINERS_LOCALSTACK_REGION}'))
         .build()
 
-s3.createBucket(b -> b.bucket("${aws.s3.bucketNameOrArn}"))
+s3.createBucket(b -> b.bucket('${aws.s3.bucketNameOrArn}'))
 
 return s3